Break apart recorder into tasks and core modules (#71222)
