+++ .strict-typing
@@ -179,6 +179,7 @@ homeassistant.components.rdw.*
 homeassistant.components.recollect_waste.*
 homeassistant.components.recorder
 homeassistant.components.recorder.const
+homeassistant.components.recorder.core
 homeassistant.components.recorder.backup
 homeassistant.components.recorder.executor
 homeassistant.components.recorder.history
@@ -189,6 +190,7 @@ homeassistant.components.recorder.repack
 homeassistant.components.recorder.run_history
 homeassistant.components.recorder.statistics
 homeassistant.components.recorder.system_health
+homeassistant.components.recorder.tasks
 homeassistant.components.recorder.util
 homeassistant.components.recorder.websocket_api
 homeassistant.components.remote.*
