+++ homeassistant/components/recorder/const.py
@@ -28,3 +28,12 @@ DB_WORKER_PREFIX = "DbWorker"
 JSON_DUMP: Final = partial(json.dumps, cls=JSONEncoder, separators=(",", ":"))
 
 ALL_DOMAIN_EXCLUDE_ATTRS = {ATTR_ATTRIBUTION, ATTR_RESTORED, ATTR_SUPPORTED_FEATURES}
+
+ATTR_KEEP_DAYS = "keep_days"
+ATTR_REPACK = "repack"
+ATTR_APPLY_FILTER = "apply_filter"
+
+KEEPALIVE_TIME = 30
+
+
+EXCLUDE_ATTRIBUTES = f"{DOMAIN}_exclude_attributes_by_domain"
