+++ mypy.ini
@@ -1732,6 +1732,17 @@ no_implicit_optional = true
 warn_return_any = true
 warn_unreachable = true
 
+[mypy-homeassistant.components.recorder.core]
+check_untyped_defs = true
+disallow_incomplete_defs = true
+disallow_subclassing_any = true
+disallow_untyped_calls = true
+disallow_untyped_decorators = true
+disallow_untyped_defs = true
+no_implicit_optional = true
+warn_return_any = true
+warn_unreachable = true
+
 [mypy-homeassistant.components.recorder.backup]
 check_untyped_defs = true
 disallow_incomplete_defs = true
@@ -1832,6 +1843,17 @@ warn_return_any = true
 warn_unreachable = true
 
 [mypy-homeassistant.components.recorder.system_health]
+check_untyped_defs = true
+disallow_incomplete_defs = true
+disallow_subclassing_any = true
+disallow_untyped_calls = true
+disallow_untyped_decorators = true
+disallow_untyped_defs = true
+no_implicit_optional = true
+warn_return_any = true
+warn_unreachable = true
+
+[mypy-homeassistant.components.recorder.tasks]
 check_untyped_defs = true
 disallow_incomplete_defs = true
 disallow_subclassing_any = true
