+++ tests/components/recorder/test_history.py
@@ -52,7 +52,7 @@ async def _async_get_states(
 def _add_db_entries(
     hass: ha.HomeAssistant, point: datetime, entity_ids: list[str]
 ) -> None:
-    with recorder.session_scope(hass=hass) as session:
+    with session_scope(hass=hass) as session:
         for idx, entity_id in enumerate(entity_ids):
             session.add(
                 Events(
@@ -87,7 +87,9 @@ def _setup_get_states(hass):
     """Set up for testing get_states."""
     states = []
     now = dt_util.utcnow()
-    with patch("homeassistant.components.recorder.dt_util.utcnow", return_value=now):
+    with patch(
+        "homeassistant.components.recorder.core.dt_util.utcnow", return_value=now
+    ):
         for i in range(5):
             state = ha.State(
                 f"test.point_in_time_{i % 5}",
@@ -102,7 +104,9 @@ def _setup_get_states(hass):
         wait_recording_done(hass)
 
     future = now + timedelta(seconds=1)
-    with patch("homeassistant.components.recorder.dt_util.utcnow", return_value=future):
+    with patch(
+        "homeassistant.components.recorder.core.dt_util.utcnow", return_value=future
+    ):
         for i in range(5):
             state = ha.State(
                 f"test.point_in_time_{i % 5}",
@@ -122,7 +126,7 @@ def test_get_full_significant_states_with_session_entity_no_matches(hass_recorde
     hass = hass_recorder()
     now = dt_util.utcnow()
     time_before_recorder_ran = now - timedelta(days=1000)
-    with recorder.session_scope(hass=hass) as session:
+    with session_scope(hass=hass) as session:
         assert (
             history.get_full_significant_states_with_session(
                 hass, session, time_before_recorder_ran, now, entity_ids=["demo.id"]
@@ -148,7 +152,7 @@ def test_significant_states_with_session_entity_minimal_response_no_matches(
     hass = hass_recorder()
     now = dt_util.utcnow()
     time_before_recorder_ran = now - timedelta(days=1000)
-    with recorder.session_scope(hass=hass) as session:
+    with session_scope(hass=hass) as session:
         assert (
             history.get_significant_states_with_session(
                 hass,
@@ -197,11 +201,15 @@ def test_state_changes_during_period(hass_recorder, attributes, no_attributes, l
     point = start + timedelta(seconds=1)
     end = point + timedelta(seconds=1)
 
-    with patch("homeassistant.components.recorder.dt_util.utcnow", return_value=start):
+    with patch(
+        "homeassistant.components.recorder.core.dt_util.utcnow", return_value=start
+    ):
         set_state("idle")
         set_state("YouTube")
 
-    with patch("homeassistant.components.recorder.dt_util.utcnow", return_value=point):
+    with patch(
+        "homeassistant.components.recorder.core.dt_util.utcnow", return_value=point
+    ):
         states = [
             set_state("idle"),
             set_state("Netflix"),
@@ -209,7 +217,9 @@ def test_state_changes_during_period(hass_recorder, attributes, no_attributes, l
             set_state("YouTube"),
         ]
 
-    with patch("homeassistant.components.recorder.dt_util.utcnow", return_value=end):
+    with patch(
+        "homeassistant.components.recorder.core.dt_util.utcnow", return_value=end
+    ):
         set_state("Netflix")
         set_state("Plex")
 
@@ -235,11 +245,15 @@ def test_state_changes_during_period_descending(hass_recorder):
     point = start + timedelta(seconds=1)
     end = point + timedelta(seconds=1)
 
-    with patch("homeassistant.components.recorder.dt_util.utcnow", return_value=start):
+    with patch(
+        "homeassistant.components.recorder.core.dt_util.utcnow", return_value=start
+    ):
         set_state("idle")
         set_state("YouTube")
 
-    with patch("homeassistant.components.recorder.dt_util.utcnow", return_value=point):
+    with patch(
+        "homeassistant.components.recorder.core.dt_util.utcnow", return_value=point
+    ):
         states = [
             set_state("idle"),
             set_state("Netflix"),
@@ -247,7 +261,9 @@ def test_state_changes_during_period_descending(hass_recorder):
             set_state("YouTube"),
         ]
 
-    with patch("homeassistant.components.recorder.dt_util.utcnow", return_value=end):
+    with patch(
+        "homeassistant.components.recorder.core.dt_util.utcnow", return_value=end
+    ):
         set_state("Netflix")
         set_state("Plex")
 
@@ -277,14 +293,20 @@ def test_get_last_state_changes(hass_recorder):
     point = start + timedelta(minutes=1)
     point2 = point + timedelta(minutes=1)
 
-    with patch("homeassistant.components.recorder.dt_util.utcnow", return_value=start):
+    with patch(
+        "homeassistant.components.recorder.core.dt_util.utcnow", return_value=start
+    ):
         set_state("1")
 
     states = []
-    with patch("homeassistant.components.recorder.dt_util.utcnow", return_value=point):
+    with patch(
+        "homeassistant.components.recorder.core.dt_util.utcnow", return_value=point
+    ):
         states.append(set_state("2"))
 
-    with patch("homeassistant.components.recorder.dt_util.utcnow", return_value=point2):
+    with patch(
+        "homeassistant.components.recorder.core.dt_util.utcnow", return_value=point2
+    ):
         states.append(set_state("3"))
 
     hist = history.get_last_state_changes(hass, 2, entity_id)
@@ -310,10 +332,14 @@ def test_ensure_state_can_be_copied(hass_recorder):
     start = dt_util.utcnow() - timedelta(minutes=2)
     point = start + timedelta(minutes=1)
 
-    with patch("homeassistant.components.recorder.dt_util.utcnow", return_value=start):
+    with patch(
+        "homeassistant.components.recorder.core.dt_util.utcnow", return_value=start
+    ):
         set_state("1")
 
-    with patch("homeassistant.components.recorder.dt_util.utcnow", return_value=point):
+    with patch(
+        "homeassistant.components.recorder.core.dt_util.utcnow", return_value=point
+    ):
         set_state("2")
 
     hist = history.get_last_state_changes(hass, 2, entity_id)
@@ -486,23 +512,28 @@ def test_get_significant_states_only(hass_recorder):
         points.append(start + timedelta(minutes=i))
 
     states = []
-    with patch("homeassistant.components.recorder.dt_util.utcnow", return_value=start):
+    with patch(
+        "homeassistant.components.recorder.core.dt_util.utcnow", return_value=start
+    ):
         set_state("123", attributes={"attribute": 10.64})
 
     with patch(
-        "homeassistant.components.recorder.dt_util.utcnow", return_value=points[0]
+        "homeassistant.components.recorder.core.dt_util.utcnow",
+        return_value=points[0],
     ):
         # Attributes are different, state not
         states.append(set_state("123", attributes={"attribute": 21.42}))
 
     with patch(
-        "homeassistant.components.recorder.dt_util.utcnow", return_value=points[1]
+        "homeassistant.components.recorder.core.dt_util.utcnow",
+        return_value=points[1],
     ):
         # state is different, attributes not
         states.append(set_state("32", attributes={"attribute": 21.42}))
 
     with patch(
-        "homeassistant.components.recorder.dt_util.utcnow", return_value=points[2]
+        "homeassistant.components.recorder.core.dt_util.utcnow",
+        return_value=points[2],
     ):
         # everything is different
         states.append(set_state("412", attributes={"attribute": 54.23}))
@@ -547,7 +578,9 @@ def record_states(hass):
     four = three + timedelta(seconds=1)
 
     states = {therm: [], therm2: [], mp: [], mp2: [], mp3: [], script_c: []}
-    with patch("homeassistant.components.recorder.dt_util.utcnow", return_value=one):
+    with patch(
+        "homeassistant.components.recorder.core.dt_util.utcnow", return_value=one
+    ):
         states[mp].append(
             set_state(mp, "idle", attributes={"media_title": str(sentinel.mt1)})
         )
@@ -564,7 +597,9 @@ def record_states(hass):
             set_state(therm, 20, attributes={"current_temperature": 19.5})
         )
 
-    with patch("homeassistant.components.recorder.dt_util.utcnow", return_value=two):
+    with patch(
+        "homeassistant.components.recorder.core.dt_util.utcnow", return_value=two
+    ):
         # This state will be skipped only different in time
         set_state(mp, "YouTube", attributes={"media_title": str(sentinel.mt3)})
         # This state will be skipped because domain is excluded
@@ -579,7 +614,9 @@ def record_states(hass):
             set_state(therm2, 20, attributes={"current_temperature": 19})
         )
 
-    with patch("homeassistant.components.recorder.dt_util.utcnow", return_value=three):
+    with patch(
+        "homeassistant.components.recorder.core.dt_util.utcnow", return_value=three
+    ):
         states[mp].append(
             set_state(mp, "Netflix", attributes={"media_title": str(sentinel.mt4)})
         )
