+++ tests/components/recorder/test_websocket_api.py
@@ -323,9 +323,10 @@ async def test_recorder_info_migration_queue_exhausted(hass, hass_ws_client):
     with patch("homeassistant.components.recorder.ALLOW_IN_MEMORY_DB", True), patch(
         "homeassistant.components.recorder.Recorder.async_periodic_statistics"
     ), patch(
-        "homeassistant.components.recorder.create_engine", new=create_engine_test
+        "homeassistant.components.recorder.core.create_engine",
+        new=create_engine_test,
     ), patch.object(
-        recorder, "MAX_QUEUE_BACKLOG", 1
+        recorder.core, "MAX_QUEUE_BACKLOG", 1
     ), patch(
         "homeassistant.components.recorder.migration.migrate_schema",
         wraps=stalled_migration,
@@ -384,7 +385,7 @@ async def test_backup_start_timeout(
     # Ensure there are no queued events
     await async_wait_recording_done(hass)
 
-    with patch.object(recorder, "DB_LOCK_TIMEOUT", 0):
+    with patch.object(recorder.core, "DB_LOCK_TIMEOUT", 0):
         try:
             await client.send_json({"id": 1, "type": "backup/start"})
             response = await client.receive_json()
